stage 1
add ATMRepository test

+++ src/main/java/org/arkaan/simpleatm/datamodel/Customer.java
@@ -5,20 +5,20 @@ import java.util.List;
 
 public class Customer {
     private final String name;
-    private double balance;
+    private Integer balance;
     private final List<Transaction> transactionList;
 
-    Customer(String name, double balance) {
+    Customer(String name, Integer balance) {
         transactionList = new ArrayList<>();
         this.name = name;
         this.balance = balance;
     }
 
-    void addBalance(double amount) {
+    void addBalance(Integer amount) {
         balance += amount;
     }
 
-    void reduceBalance(double amount) {
+    void reduceBalance(Integer amount) {
         balance -= amount;
     }
 
@@ -26,7 +26,7 @@ public class Customer {
         return name;
     }
 
-    double getBalance() {
+    Integer getBalance() {
         return balance;
     }
